requirements.txt
langchain~=0.3.27
langchain-community~=0.3.27
langgraph
transformers>=4.40.0
sentencepiece>=0.1.99
huggingface_hub
langchain-huggingface
gradio~=5.41.1
arxiv
PyPDF2~=3.0.1
pytest~=8.4.1
torch
sentence-transformers
faiss-cpu
keybert~=0.9.0
python-dotenv~=1.1.1
requests~=2.32.4
feedparser~=6.0.11
streamlit